requirements.txt
google-api-core==2.11.0
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
packaging==21.3
python-dateutil==2.8.2
protobuf==4.22.1

